requirements.txt
pulsar-client~=3.1.0
fastavro~=0.24.0
